Update scripts for train/test split
Previously, `format_inputs` in preprocessing.py was not set up to take
in the actual ratios (targets) and filter those along with the reviews
(e.g. so the ratio corresponding to a filtered review was also
filtered). This was discovered in trying to generate a train/test
split, and `format_inputs` as well as its test was updated to reflect it.