Alter preprocessing to allow for fair comparison
Prior preprocessing didn't allow for any easy way to feed the exact same
inputs into the neural-network and non-neural-network models. The
changes made here save a version of the filtered reviews/ratios much
earlier so that the same version can be fed into all models and fair
comparison can be made between the neural-network models and others.
There is still work to be done to complete this process (e.g. reworking
the net architecture to account for this), but this is a big step
torwards fair comparison.