Rework net inputs to pad sequences
This commit is a natural extension of the previous commit. By reworking
the net inputs to pad the sequences rather than discarding sequences that
didn't meet a specific length, the inputs to the net will line up with
the inputs to non-net models. Padding the sequences further meant that
the `idx_word_dct` could no longer have 0 denoting the UNK symbol, since
sequences are padded with 0.