Add more terminal runtime options to keras model
I have also altered the `build` method of the KerasSeq2NoSeq. With
`mean_squared_error` loss, the model is meant to use a `linear`
activation function in the output later, and with `binary_crossentropy`
a `softmax` activation function. There isn't meant to be other options
with these losses, and now there isn't.