Move tsne plotting into a class, fix tests for it
Note that before this PR, a number of tests were broken (and some still
are). This PR fixes tests associated with the newly built TSNEPlotter,
as well as all tests in `test_tsne_plotting` **except**
`test_gen_tsne_embedding`, because that will be moved into a class in a
subsequent PR. This leaves one other test broken - the one in
`test_eval_utils`.